organized csharp and go codes
