+++ go/algorithms/binarySearch.go
